+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ------
 
+* Generate `Mirage.Factory` attributes from a JSON Schema.
 * Generate `DS.hasMany` and `DS.belongsTo` relationships from JSON Schema.
 
 0.0.1